+++ packages/runtime/container-runtime/src/summary/summarizer.ts
@@ -4,21 +4,12 @@
  */
 
 import { TypedEventEmitter } from "@fluid-internal/client-utils";
-import { LoaderHeader } from "@fluidframework/container-definitions/internal";
-import type { ILoader } from "@fluidframework/container-definitions/internal";
 import type {
 	ISummarizerEvents,
 	SummarizerStopReason,
 } from "@fluidframework/container-runtime-definitions/internal";
-import {
-	IFluidHandleContext,
-	type FluidObject,
-	type IRequest,
-	type IResponse,
-} from "@fluidframework/core-interfaces/internal";
+import { IFluidHandleContext } from "@fluidframework/core-interfaces/internal";
 import { Deferred } from "@fluidframework/core-utils/internal";
-import { DriverHeader } from "@fluidframework/driver-definitions/internal";
-import { responseToException } from "@fluidframework/runtime-utils/internal";
 import {
 	IFluidErrorBase,
 	ITelemetryLoggerExt,
@@ -31,7 +22,6 @@ import {
 // eslint-disable-next-line import/no-deprecated
 import { ICancellableSummarizerController } from "./runWhileConnectedCoordinator.js";
 import { RunningSummarizer } from "./runningSummarizer.js";
-import { summarizerClientType } from "./summarizerClientElection.js";
 import { SummarizeHeuristicData } from "./summarizerHeuristics.js";
 import {
 	EnqueueSummarizeResult,
@@ -48,165 +38,16 @@ import {
 	ISummarizerRuntime,
 	ISummarizingWarning,
 	type IRetriableFailureError,
+	type ISummaryConfiguration,
 } from "./summarizerTypes.js";
 import { SummaryCollection } from "./summaryCollection.js";
 import { SummarizeResultBuilder } from "./summaryGenerator.js";
-
-const summarizingError = "summarizingError";
-
-/**
- * @legacy
- * @alpha
- */
-export interface ISummaryBaseConfiguration {
-	/**
-	 * Delay before first attempt to spawn summarizing container.
-	 */
-	initialSummarizerDelayMs: number;
-
-	/**
-	 * Defines the maximum allowed time to wait for a pending summary ack.
-	 * The maximum amount of time client will wait for a summarize is the minimum of
-	 * maxSummarizeAckWaitTime (currently 3 * 60 * 1000) and maxAckWaitTime.
-	 */
-	maxAckWaitTime: number;
-	/**
-	 * Defines the maximum number of Ops in between Summaries that can be
-	 * allowed before forcibly electing a new summarizer client.
-	 */
-	maxOpsSinceLastSummary: number;
-}
-
-/**
- * @legacy
- * @alpha
- */
-export interface ISummaryConfigurationHeuristics extends ISummaryBaseConfiguration {
-	state: "enabled";
-	/**
-	 * Defines the maximum allowed time, since the last received Ack, before running the summary
-	 * with reason maxTime.
-	 * For example, say we receive ops one by one just before the idle time is triggered.
-	 * In this case, we still want to run a summary since it's been a while since the last summary.
-	 */
-	maxTime: number;
-	/**
-	 * Defines the maximum number of Ops, since the last received Ack, that can be allowed
-	 * before running the summary with reason maxOps.
-	 */
-	maxOps: number;
-	/**
-	 * Defines the minimum number of Ops, since the last received Ack, that can be allowed
-	 * before running the last summary.
-	 */
-	minOpsForLastSummaryAttempt: number;
-	/**
-	 * Defines the lower boundary for the allowed time in between summarizations.
-	 * Pairs with maxIdleTime to form a range.
-	 * For example, if we only receive 1 op, we don't want to have the same idle time as say 100 ops.
-	 * Based on the boundaries we set in minIdleTime and maxIdleTime, the idle time will change
-	 * linearly depending on the number of ops we receive.
-	 */
-	minIdleTime: number;
-	/**
-	 * Defines the upper boundary for the allowed time in between summarizations.
-	 * Pairs with minIdleTime to form a range.
-	 * For example, if we only receive 1 op, we don't want to have the same idle time as say 100 ops.
-	 * Based on the boundaries we set in minIdleTime and maxIdleTime, the idle time will change
-	 * linearly depending on the number of ops we receive.
-	 */
-	maxIdleTime: number;
-	/**
-	 * Runtime op weight to use in heuristic summarizing.
-	 * This number is a multiplier on the number of runtime ops we process when running summarize heuristics.
-	 * For example: (multiplier) * (number of runtime ops) = weighted number of runtime ops
-	 */
-	runtimeOpWeight: number;
-	/**
-	 * Non-runtime op weight to use in heuristic summarizing
-	 * This number is a multiplier on the number of non-runtime ops we process when running summarize heuristics.
-	 * For example: (multiplier) * (number of non-runtime ops) = weighted number of non-runtime ops
-	 */
-	nonRuntimeOpWeight: number;
-
-	/**
-	 * Number of ops since last summary needed before a non-runtime op can trigger running summary heuristics.
-	 *
-	 * Note: Any runtime ops sent before the threshold is reached will trigger heuristics normally.
-	 * This threshold ONLY applies to non-runtime ops triggering summaries.
-	 *
-	 * For example: Say the threshold is 20. Sending 19 non-runtime ops will not trigger any heuristic checks.
-	 * Sending the 20th non-runtime op will trigger the heuristic checks for summarizing.
-	 */
-	nonRuntimeHeuristicThreshold?: number;
-}
-
-/**
- * @legacy
- * @alpha
- */
-export interface ISummaryConfigurationDisableSummarizer {
-	state: "disabled";
-}
-
-/**
- * @legacy
- * @alpha
- */
-export interface ISummaryConfigurationDisableHeuristics extends ISummaryBaseConfiguration {
-	state: "disableHeuristics";
-}
-
-/**
- * @legacy
- * @alpha
- */
-export type ISummaryConfiguration =
-	| ISummaryConfigurationDisableSummarizer
-	| ISummaryConfigurationDisableHeuristics
-	| ISummaryConfigurationHeuristics;
-
-export function isSummariesDisabled(
-	config: ISummaryConfiguration,
-): config is ISummaryConfigurationDisableSummarizer {
-	return config.state === "disabled";
-}
-
-/**
- * @legacy
- * @alpha
- */
-export const DefaultSummaryConfiguration: ISummaryConfiguration = {
-	state: "enabled",
-
-	minIdleTime: 0,
-
-	maxIdleTime: 30 * 1000, // 30 secs.
-
-	maxTime: 60 * 1000, // 1 min.
-
-	maxOps: 100, // Summarize if 100 weighted ops received since last snapshot.
-
-	minOpsForLastSummaryAttempt: 10,
-
-	maxAckWaitTime: 3 * 60 * 1000, // 3 mins.
-
-	maxOpsSinceLastSummary: 7000,
-
-	initialSummarizerDelayMs: 5 * 1000, // 5 secs.
-
-	nonRuntimeOpWeight: 0.1,
-
-	runtimeOpWeight: 1,
-
-	nonRuntimeHeuristicThreshold: 20,
-};
 
 export class SummarizingWarning
 	extends LoggingError
 	implements ISummarizingWarning, IFluidErrorBase
 {
-	readonly errorType = summarizingError;
+	readonly errorType = "summarizingError";
 	readonly canRetry = true;
 
 	constructor(
@@ -227,74 +68,10 @@ export class SummarizingWarning
 	}
 }
 
-export const summarizerRequestUrl = "_summarizer";
-
 export const createSummarizingWarning = (
 	errorMessage: string,
 	logged: boolean,
 ): SummarizingWarning => new SummarizingWarning(errorMessage, logged);
-
-/**
- * Returns a function that will create and retrieve a Summarizer.
- */
-export function formCreateSummarizerFn(loader: ILoader): () => Promise<ISummarizer> {
-	return async () => {
-		const request: IRequest = {
-			headers: {
-				[LoaderHeader.cache]: false,
-				[LoaderHeader.clientDetails]: {
-					capabilities: { interactive: false },
-					type: summarizerClientType,
-				},
-				[DriverHeader.summarizingClient]: true,
-				[LoaderHeader.reconnect]: false,
-			},
-			url: `/${summarizerRequestUrl}`,
-		};
-
-		const resolvedContainer = await loader.resolve(request);
-		let fluidObject: FluidObject<ISummarizer> | undefined;
-
-		// Older containers may not have the "getEntryPoint" API
-		// ! This check will need to stay until LTS of loader moves past 2.0.0-internal.7.0.0
-		if (resolvedContainer.getEntryPoint === undefined) {
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-explicit-any
-			const response = (await (resolvedContainer as any).request({
-				url: `/${summarizerRequestUrl}`,
-			})) as IResponse;
-			if (response.status !== 200 || response.mimeType !== "fluid/object") {
-				throw responseToException(response, request);
-			}
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-			fluidObject = response.value;
-		} else {
-			fluidObject = await resolvedContainer.getEntryPoint();
-		}
-
-		if (fluidObject?.ISummarizer === undefined) {
-			throw new UsageError("Fluid object does not implement ISummarizer");
-		}
-		return fluidObject.ISummarizer;
-	};
-}
-
-export function validateSummaryHeuristicConfiguration(
-	configuration: ISummaryConfigurationHeuristics,
-): void {
-	// eslint-disable-next-line no-restricted-syntax
-	for (const prop in configuration) {
-		if (typeof configuration[prop] === "number" && configuration[prop] < 0) {
-			throw new UsageError(
-				`Summary heuristic configuration property "${prop}" cannot be less than 0`,
-			);
-		}
-	}
-	if (configuration.minIdleTime > configuration.maxIdleTime) {
-		throw new UsageError(
-			`"minIdleTime" [${configuration.minIdleTime}] cannot be greater than "maxIdleTime" [${configuration.maxIdleTime}]`,
-		);
-	}
-}
 
 /**
  * Summarizer is responsible for coordinating when to generate and send summaries.
