+++ packages/runtime/container-runtime/src/summary/summarizerTypes.ts
@@ -27,7 +27,6 @@ import {
 	ITelemetryLoggerPropertyBag,
 } from "@fluidframework/telemetry-utils/internal";
 
-import type { ISummaryConfigurationHeuristics } from "./summarizer.js";
 import {
 	ISummaryAckMessage,
 	ISummaryNackMessage,
@@ -772,3 +771,115 @@ export interface ISummarizeRunnerTelemetry extends ITelemetryLoggerPropertyBag {
 	 */
 	summarizerSuccessfulAttempts: () => number;
 }
+
+/**
+ * @legacy
+ * @alpha
+ */
+export interface ISummaryBaseConfiguration {
+	/**
+	 * Delay before first attempt to spawn summarizing container.
+	 */
+	initialSummarizerDelayMs: number;
+
+	/**
+	 * Defines the maximum allowed time to wait for a pending summary ack.
+	 * The maximum amount of time client will wait for a summarize is the minimum of
+	 * maxSummarizeAckWaitTime (currently 3 * 60 * 1000) and maxAckWaitTime.
+	 */
+	maxAckWaitTime: number;
+	/**
+	 * Defines the maximum number of Ops in between Summaries that can be
+	 * allowed before forcibly electing a new summarizer client.
+	 */
+	maxOpsSinceLastSummary: number;
+}
+
+/**
+ * @legacy
+ * @alpha
+ */
+export interface ISummaryConfigurationHeuristics extends ISummaryBaseConfiguration {
+	state: "enabled";
+	/**
+	 * Defines the maximum allowed time, since the last received Ack, before running the summary
+	 * with reason maxTime.
+	 * For example, say we receive ops one by one just before the idle time is triggered.
+	 * In this case, we still want to run a summary since it's been a while since the last summary.
+	 */
+	maxTime: number;
+	/**
+	 * Defines the maximum number of Ops, since the last received Ack, that can be allowed
+	 * before running the summary with reason maxOps.
+	 */
+	maxOps: number;
+	/**
+	 * Defines the minimum number of Ops, since the last received Ack, that can be allowed
+	 * before running the last summary.
+	 */
+	minOpsForLastSummaryAttempt: number;
+	/**
+	 * Defines the lower boundary for the allowed time in between summarizations.
+	 * Pairs with maxIdleTime to form a range.
+	 * For example, if we only receive 1 op, we don't want to have the same idle time as say 100 ops.
+	 * Based on the boundaries we set in minIdleTime and maxIdleTime, the idle time will change
+	 * linearly depending on the number of ops we receive.
+	 */
+	minIdleTime: number;
+	/**
+	 * Defines the upper boundary for the allowed time in between summarizations.
+	 * Pairs with minIdleTime to form a range.
+	 * For example, if we only receive 1 op, we don't want to have the same idle time as say 100 ops.
+	 * Based on the boundaries we set in minIdleTime and maxIdleTime, the idle time will change
+	 * linearly depending on the number of ops we receive.
+	 */
+	maxIdleTime: number;
+	/**
+	 * Runtime op weight to use in heuristic summarizing.
+	 * This number is a multiplier on the number of runtime ops we process when running summarize heuristics.
+	 * For example: (multiplier) * (number of runtime ops) = weighted number of runtime ops
+	 */
+	runtimeOpWeight: number;
+	/**
+	 * Non-runtime op weight to use in heuristic summarizing
+	 * This number is a multiplier on the number of non-runtime ops we process when running summarize heuristics.
+	 * For example: (multiplier) * (number of non-runtime ops) = weighted number of non-runtime ops
+	 */
+	nonRuntimeOpWeight: number;
+
+	/**
+	 * Number of ops since last summary needed before a non-runtime op can trigger running summary heuristics.
+	 *
+	 * Note: Any runtime ops sent before the threshold is reached will trigger heuristics normally.
+	 * This threshold ONLY applies to non-runtime ops triggering summaries.
+	 *
+	 * For example: Say the threshold is 20. Sending 19 non-runtime ops will not trigger any heuristic checks.
+	 * Sending the 20th non-runtime op will trigger the heuristic checks for summarizing.
+	 */
+	nonRuntimeHeuristicThreshold?: number;
+}
+
+/**
+ * @legacy
+ * @alpha
+ */
+export interface ISummaryConfigurationDisableSummarizer {
+	state: "disabled";
+}
+
+/**
+ * @legacy
+ * @alpha
+ */
+export interface ISummaryConfigurationDisableHeuristics extends ISummaryBaseConfiguration {
+	state: "disableHeuristics";
+}
+
+/**
+ * @legacy
+ * @alpha
+ */
+export type ISummaryConfiguration =
+	| ISummaryConfigurationDisableSummarizer
+	| ISummaryConfigurationDisableHeuristics
+	| ISummaryConfigurationHeuristics;
